Mars ISRU 1000 Update
RealPlumes moved to RO repo.

Removing old test craft file, as it used parts that no longer exist.
I've standardized the part names and changed the model used to one that
exists again.

2 new test craft, on that replaced the deleted one. Take a few hundred
days to fill, doesn't have its own landing capability (you could add it,
or just HyperEdit it to the surface of Mars for testing)

Second new craft uses the new 1000 class units, capable of producing a
tonne of fuel per day if run continuously. The 1000 class ISRU is
oversized for this test craft, so it only takes about a month to fuel
up. This craft is capable of direct Mars landing and generating lift.
The chute is meant for Earth return, not Mars use. RTGs including here
are for testing and are unrealistically massed. As built, is capable of
reaching Mars orbit and nearly escaping.

Operating these craft (for me at least) requires raising the boiling
points for the cryogens. If you don't know how to do that, here is my
file for that:

https://www.dropbox.com/sh/vf0nngt4ne0lulf/AAAIUx1tpicGCE1NDe76eNXJa?dl=0